requirements.txt
Django==3.2
djangorestframework==3.12.3
gunicorn==20.1.0
djangorestframework-simplejwt==4.8.0
djoser
dj-database-url==1.2.0
pymysql==1.0.2
python-dotenv==0.21.0

PyJWT==2.6.0
psycopg2-binary==2.9.5
social-auth-app-django==4.0.0
dj-rest-auth==2.2.5
django-allauth==0.51.0
drf-yasg==1.21.4
Pillow==9.3
pandas==1.5.3
openpyxl==3.1.1